requirements.txt
streamlit
openai
pandas
faiss-cpu
numpy
matplotlib
datasets
langchain-openai
numpy
pandas
ragas
plotly
